Move CPU usage related functions from profile-data to cpu file
